+++ Mouse-Bounder/BoundType.cs
@@ -1,0 +1,9 @@
+﻿namespace Mouse_Bounder
+{
+    public enum BoundType
+    {
+        MouseEvent = 0,
+        Polling = 1,
+        Both = 2
+    }
+}